Fix variable type and add fontsizing

+++ src/main/java/ch/fhnw/ip5/digitalfontclassification/Point.java
@@ -1,30 +1,30 @@
 package ch.fhnw.ip5.digitalfontclassification;
 
 public class Point {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
     public Point() {
         this.x = 0.0f;
         this.y = 0.0f;
     }
 
-    public Point(float x, float y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
-    public float getY() {
+    public double getY() {
         return y;
     }
